tests/wpt/mozilla/tests/mozilla/inline-event-listener-panic.html

<!doctype html>
<meta charset="utf-8">
<title>Getting the value of an inline event handler shouldn't panic.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
test(function() {
  var e = document.body;
  e.setAttribute("onclick", "console.log('x')");
  assert_not_equals(e.onclick, null);
  assert_not_equals(e.onclick, null);
});
</script>
</body>
